+++ GeometryCore/GeometryCore/Vector.h
@@ -9,7 +9,7 @@ namespace Geometry
 	{
 	public:
 		Vector2D() noexcept;
-		Vector2D(double x, double y) noexcept;
+		Vector2D(float x, float y) noexcept;
 
 		Vector2D(const Vector2D& other) noexcept;
 		Vector2D(Vector2D&& other) noexcept;
@@ -18,27 +18,27 @@ namespace Geometry
 		~Vector2D() noexcept;
 
 		bool operator==(const Vector2D& other) const noexcept;
-		double& operator[](int index) noexcept;
-		double operator[](int index) const noexcept;
+		float& operator[](int index) noexcept;
+		float operator[](int index) const noexcept;
 		Vector2D operator+(const Vector2D& other) const noexcept;
 		Vector2D operator-(const Vector2D& other) const noexcept;
-		Vector2D operator/(double n) const noexcept;
+		Vector2D operator/(float n) const noexcept;
 		void operator+=(const Vector2D& other) noexcept;
 
-		double x() const noexcept;
-		void setX(double x) noexcept;
-		double y() const noexcept;
-		void setY(double y) noexcept;
+		float x() const noexcept;
+		void setX(float x) noexcept;
+		float y() const noexcept;
+		void setY(float y) noexcept;
 
-		double r() const noexcept;
-		void setR(double r) noexcept;
-		double g() const noexcept;
-		void setG(double g) noexcept;
+		float r() const noexcept;
+		void setR(float r) noexcept;
+		float g() const noexcept;
+		void setG(float g) noexcept;
 
 		Vec4Impl* __internal_getPimpl() const noexcept;
 
 		static Vector2D normalize(const Vector2D& vec) noexcept;
-		static double dot(const Vector2D& firstVec, const Vector2D& secondVec) noexcept;
+		static float dot(const Vector2D& firstVec, const Vector2D& secondVec) noexcept;
 		static Vector2D cross(const Vector2D& firstVec, const Vector2D& secondVec) noexcept;
 	protected:
 		Vector2D(Vec4Impl&& other) noexcept;
@@ -53,7 +53,7 @@ namespace Geometry
 	{
 	public:
 		Vector3D() noexcept = default;
-		Vector3D(double x, double y, double z) noexcept;
+		Vector3D(float x, float y, float z) noexcept;
 		Vector3D(Vector2D&& other) noexcept;
 
 		Vector3D(const Vector3D& other) noexcept;
@@ -63,15 +63,15 @@ namespace Geometry
 
 		Vector3D operator+(const Vector3D& other) const noexcept;
 		Vector3D operator-(const Vector3D& other) const noexcept;
-		Vector3D operator/(double n) const noexcept;
+		Vector3D operator/(float n) const noexcept;
 
-		double z() const noexcept;
-		void setZ(double z) noexcept;
-		double b() const noexcept;
-		void setB(double b) noexcept;
+		float z() const noexcept;
+		void setZ(float z) noexcept;
+		float b() const noexcept;
+		void setB(float b) noexcept;
 
 		static Vector3D normalize(const Vector3D& vec) noexcept;
-		static double dot(const Vector3D& firstVec, const Vector3D& secondVec) noexcept;
+		static float dot(const Vector3D& firstVec, const Vector3D& secondVec) noexcept;
 		static Vector3D cross(const Vector3D& firstVec, const Vector3D& secondVec) noexcept;
 		static Vector3D project(const Vector3D& obj, const Matrix4D& model, const Matrix4D& proj, const Vector4D& viewport) noexcept;
 		static Vector3D unProject(const Vector3D& win, const Matrix4D& model, const Matrix4D& proj, const Vector4D& viewport) noexcept;
@@ -84,7 +84,7 @@ namespace Geometry
 	{
 	public:
 		Vector4D() noexcept = default;
-		Vector4D(double x, double y, double z, double w) noexcept;
+		Vector4D(float x, float y, float z, float w) noexcept;
 		Vector4D(Vector3D&& other) noexcept;
 
 		Vector4D(const Vector4D& other) noexcept;
@@ -94,18 +94,18 @@ namespace Geometry
 
 		Vector4D operator+(const Vector4D& other) const noexcept;
 		Vector4D operator-(const Vector4D& other) const noexcept;
-		Vector4D operator/(double n) const noexcept;
+		Vector4D operator/(float n) const noexcept;
 		Vector4D operator*(const Matrix4D& matrix) const noexcept;
 
 		Vector3D getVec3() const noexcept;
 
-		double w() const noexcept;
-		void setW(double w) noexcept;
-		double a() const noexcept;
-		void setA(double b) noexcept;
+		float w() const noexcept;
+		void setW(float w) noexcept;
+		float a() const noexcept;
+		void setA(float b) noexcept;
 
 		static Vector4D normalize(const Vector4D& vec) noexcept;
-		static double dot(const Vector4D& firstVec, const Vector4D& secondVec) noexcept;
+		static float dot(const Vector4D& firstVec, const Vector4D& secondVec) noexcept;
 		static Vector4D cross(const Vector4D& firstVec, const Vector4D& secondVec) noexcept;
 
 	protected:
